+++ _ext/eclipse-base/CHANGES.md
@@ -1,8 +1,10 @@
 # spotless-eclipse-base
 
-### Version 3.1.0-SNAPSHOT - TBD
+### Version 3.1.0 - TBD
 
-* TBD.
+* Added logging service based on SLF4J. ([#236](https://github.com/diffplug/spotless/issues/236))
+* Updated internal interfaces to support `org.eclipse.osgi` version 3.13.
+* Corrected documentation of version number usage in `gradle.properties`.
 
 ### Version 3.0.0 - July 18th 2018 ([artifact]([jcenter](https://bintray.com/diffplug/opensource/spotless-eclipse-base)))